feat: 꾸미기 요소 캔버스 컴포넌트 이동
