+++ src/feature/PentagramDecoration/common/components/style/decorationCanvas.scss
@@ -1,5 +1,5 @@
 .decoration-canvas-component {
     position: absolute;
-    border-radius: 50%;
     transform: translate(-50%, -50%);
+    clip-path: polygon(50% 0%, 98% 34%, 80% 91%, 20% 91%, 2% 34%);
 }
